[common.parallel.threadmanager] implement safe integer conversion

+++ dune/stuff/common/parallel/threadmanager.cc
@@ -52,7 +52,7 @@ size_t Dune::Stuff::ThreadManager::thread()
 void Dune::Stuff::ThreadManager::set_max_threads(const size_t count)
 {
   max_threads_ = count;
-  WITH_DUNE_FEM(Dune::Fem::ThreadManager::setMaxNumberThreads(count);)
+  WITH_DUNE_FEM(Dune::Fem::ThreadManager::setMaxNumberThreads(boost::numeric_cast<int>(count));)
 #if HAVE_EIGEN
   Eigen::setNbThreads(boost::numeric_cast<int>(count));
 #endif
